parse AlternativeContent under CT_Run
fix #769
