+++ OpenXmlFormats/Vml/SpreadsheetDrawing.cs
@@ -9,42 +9,6 @@ using NPOI.OpenXml4Net.Util;
 
 namespace NPOI.OpenXmlFormats.Vml.Spreadsheet
 {
-    public class CT_AlternateContent 
-    {
-        public string innerXml { get; set; }
-        public CT_AlternateContent() 
-        {
-        }
-        public static CT_AlternateContent Parse(XmlNode node, XmlNamespaceManager namespaceManager)
-        {
-            if (node == null)
-                return null;
-
-            var ac = new CT_AlternateContent();
-            if (string.IsNullOrEmpty(node.InnerXml))
-            {
-                return ac;
-            }
-            ac.innerXml = node.InnerXml;
-            return ac;
-        }
-        internal void Write(StreamWriter sw, string nodeName)
-        {
-            sw.Write(string.Format("<mc:{0} xmlns:mc=\"http://schemas.openxmlformats.org/markup-compatibility/2006\"", nodeName));
-            if (this.innerXml == null)
-            {
-                sw.Write(string.Format("/>", nodeName));
-            }
-            else
-            {
-                sw.Write(">");
-                sw.Write(this.innerXml);
-                sw.Write(string.Format("</mc:{0}>", nodeName));
-            }
-            
-        }
-
-    }
     [System.ComponentModel.DesignerCategory("code")]
     public class CT_ClientData
     {
